internal: refactor missing match arms diagnostics
